feat(markdown): Add daylog snippet and LuaSnip integration
This commit introduces a "daylog" snippet for Markdown files,
facilitating daily logging with sections for reflections, TODOs,
and general logs.

It also configures LuaSnip for Markdown, enabling the use of
this new snippet. A user command `InsertLog` and a keymap
`<leader>idl` are added to easily insert the "daylog" snippet.

The LuaSnip plugin is added to the Neovim configuration,
and Markdown filetype extensions are set up to load snippets
from both VSCode-style and Lua-defined snippet paths.